Move the Mysql provider service to a separate container.
Previously we were embedding our service inside the Mysql server image
so everything was inside one container. This was simpler in some ways
(e.g. because you could start it as a single isolated container) but it
unnecessarily couples us to the implementation of the Mysql container
image, which could change without warning and break us.

Now we use two separate containers, one for Mysql and one for the
provider service, and launch them as a single "project" using
docker-compose. This decouples us from the implementation of the Mysql
server container, so it can change without breaking us (as long as it
continues to honor the environment variables that configure root access,
which seems to be the correct way to configure it).

I also added a working integration test example so you can quickly test
if the service is working in your environment.