Remove "go test" from presubmit script.
AFAIK there is no way to exclude the integration tests from running, so
the script will fail if the provider service is not running in the
background (could be problematic for continuous integration test
systems not running on your workstation).

The Bazel test run covers all the unit tests, and exclude any tests marked
as "manual".