Change the flag to an environment variable.
The problem with using a flag is that this database client is used in
unit tests, which may be part of a larger test suite where not all tests
know about the given flag (so passing the flag to all tests can
automatically fail some tests). Using an env var avoids this because
tests can silently ignore environment variables that they don't care
about.